Add internal config package
The internal config package adds a ReadConfig function which reads
configuration files from the system. The configuration files are read in
order starting with the passed in path and ending with the default
system configuration path. If no configuration file is found the
compiled defaults are used. Any values not set in the configuration
files use the compiled defaults. The files are read in the following
order with the first file found being used.

Passed configuration file path, if defined
$HOME/.config/diva/config.toml
/etc/diva/config.toml
/usr/share/defaults/diva/config.toml

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>