results: improve API for recording and printing
Add name and description to the Results struct, record all tests in a
Results.Tests slice of Result structs so they can be inspected by the
caller. Print name and description of each failure when printing test
results. Combine all Add functions into a single Add function and
determine if the test Passed, Failed, or was Skipped from the arguments.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>