pkginfo: Add initial implementation of rpm data library
This patch fills out the RPM library (pkginfo) API with initial
implementations of the functions. Work still to be done is documented in
TODO comments. This patch allows users of the library to import RPMs
from an RPM repository and store them in a running redis database. Users
can also retrieve all repo information (RPMs and Files) using this API.

The getrepo.go functionality was moved to pkginfo because at least for
now it is the only consumer. It was extremely useful for getrepo
functionality to have usage of the Repo struct defined in types.go.

Signed-off-by: Matthew Johnson <matthew.johnson@intel.com>