Modify bundles to remove trailing whitespace
When a trailing space was at the end of a bundle include, it
was not matching regex matching for includes, so the line was
being considered a direct package for a bundle. To solve the
problem, trim whitespace for each line being read before
performing pattern matching. Include associated unit test for
the bug.

Signed-off-by: gabrielle beyer <gabrielle.n.beyer@intel.com>