Add build bloat check
This check allows users to detect specific size changes between bundles
from any build to another. There are two heuristics it checks:
1.) if any bundle increased by a certain size (20% default), emit a warning
2.) if any critical bundle such as os-core increased by a certain size
(10% default), throw a hard failure, because this will cause all bundles to
balloon in size.
The size limits are configurable through the cmd line, and it is possible
to get the bundle sizes for a single build by passing in only 1 version.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>