Add canary check to diva
This check takes in two full chroot directories, and reads the dynamic
symbols section of every valid ELF file to confirm there is a
`__stack_chk_fail` symbol signifying a stack canary is present. If it is
missing from either the old or new file, it is an error, and if the old
file does not exist, the new one must still contain it.

Signed-off-by: Tudor Marcu <tudor.marcu@intel.com>