Organize fetch and checks and import bundles
The architecture for DIVA is to separate the fetching
of data from the different checks. This has been
accomplished in this commit.

The repo and bundle data, when fetched, is downloaded and
imported into the databse. The update content still needs
this functionality. There are different data structs that
populate the needed information from the database, so the
check commands are not downloading any content.

Signed-off-by: gabrielle beyer <gabrielle.n.beyer@intel.com>