Faster cache updating when option passed
Instead of iterating a slice of rpm objects, convert the
slice to a map of rpm name to rpm objects. When the
upgrade cache argument is passed, it is a simple lookup to
get the currently cached rpm data. Stop using the go-rpm
library to create an object for every 'potential' new RPM,
instead compare the rpm n-v-r with the cached rpm n-v-r.
If they differ, delete the cached rpm and download the
new one.

Signed-off-by: gabrielle beyer <gabrielle.n.beyer@intel.com>