Modify pundle deletes and add bundle deletes
Add detection for bundle deletes and modify the pundle
delete detection to use the specified tag, as opposed to
latest every time. Also check that if a current bundles
repo exists, that it is the same bundleURL being fetched,
and if not delete it and clone the new one. (This was a bug,
where if a different repo URL was passed it would say
it already exists, when it was really an old url.)

Signed-off-by: gabrielle beyer <gabrielle.n.beyer@intel.com>