Download SRPM repo when fetch repo called
By default download both the binaries and SRPMs to the cache
and import both into the database. If a user wants to specify
only one RPM type, they can pass --source or --binary depending
on whichever they desire.

Add fallback functionality for the source and binary RPM url
creation. Construct the mash URL and the releases URL for
the rpm repos, then check the validity with the other info
provided from the command line. If no URL is found to be good
then return an error. Otherwise, use the good url.
Signed-off-by: gabrielle beyer <gabrielle.n.beyer@intel.com>